Add dockerfile and reorganize files
-feat: add dockerfile and imported db to docker container
-fix: index.php now only refers to login.php so you will need to log in
-chore: change main page into dashboard.php including in all pages

+++ delete.php
@@ -3,7 +3,12 @@ include "connection.php";
 
 $id = intval($_GET['id']);
 
+if ($_SESSION['role'] !== 'admin') {
+    header("Location: dashboard.php?error=1");
+    exit;
+}
+
 mysqli_query($conn, "DELETE FROM movies WHERE id = $id");
 
-header("Location: index.php?deleted=1");
+header("Location: dashboard.php?deleted=1");
 ?>
